% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tags.R
\name{set_tags}
\alias{set_tags}
\title{Change tags of a safeframe object}
\usage{
set_tags(x, ...)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A named list with variable
names in \code{x} as list names and the tags as list values. Values set to
\code{NULL} remove the tag When specifying tags, please also see
\code{default_values}.}
}
\value{
The function returns a \code{safeframe} object.
}
\description{
This function changes the \code{tags} of a \code{safeframe} object, using the same
syntax as the constructor \code{\link[=make_safeframe]{make_safeframe()}}.
}
\examples{

## create a safeframe
x <- make_safeframe(cars, speed = "Miles per hour")
tags(x)

## add new tags and fix an existing one
x <- set_tags(x, dist = "Distance")
tags(x)

## remove tags by setting them to NULL
old_tags <- tags(x)
x <- set_tags(x, speed = NULL, dist = NULL)
tags(x)

## setting tags providing a list (used to restore old tags here)
x <- set_tags(x, !!!old_tags)
tags(x)
}
\seealso{
\code{\link[=make_safeframe]{make_safeframe()}} to create a \code{safeframe} object
}
