% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_variables.R
\name{tag_variables}
\alias{tag_variables}
\title{Add tags to variables}
\usage{
tag_variables(x, tags)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}, with at least one column}

\item{tags}{A named list with variable names in \code{x} as list names and the
tags as list values. Values set to \code{NULL} remove the tag.}
}
\value{
The function returns the original object with an additional \code{"label"}
attribute on each provided variable.
}
\description{
Internal. This function will tag pre-defined variables in a
\code{data.frame} by adding a label attribute to the column. This can be used for
one or multiple variables at the same time.
}
\details{
If used several times, the previous tag is removed silently.
Only accepts known variables from the provided \code{data.frame}.
}
