% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_tags.R
\name{vars_tags}
\alias{vars_tags}
\title{Internal printing function for variables and tags}
\usage{
vars_tags(vars, tags)
}
\arguments{
\item{vars}{a \code{character} vector of variable names}

\item{tags}{a \code{character} vector of tags}
}
\description{
Internal printing function for variables and tags
}
