% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tags}
\alias{tags}
\title{Get the list of tags in a safeframe}
\usage{
tags(x, show_null = FALSE)
}
\arguments{
\item{x}{a \code{safeframe} object}

\item{show_null}{a \code{logical} indicating if the complete list of tags,
including \code{NULL} ones, should be returned; if \code{FALSE}, only tags with a
non-NULL value are returned; defaults to \code{FALSE}}
}
\value{
The function returns a named \code{list} where names indicate which column
they correspond to, and values indicate the relevant tags.
}
\description{
This function returns the list of tags identifying specific variable types
in a \code{safeframe} object.
}
\details{
tags are stored as the \code{label} attribute of the column variable.
}
\examples{

## make a safeframe
x <- make_safeframe(cars, speed = "Miles per hour")

## check non-null tags
tags(x)

## get a list of all tags, including NULL ones
tags(x, TRUE)
}
